sanity-check the Rust view of the target against the used MiniRust target
This actually found a bug in our DefaultTarget!
Also, simplify int type translation a bit.

+++ tooling/minimize/tests/pass/zst.rs
@@ -19,6 +19,6 @@ struct A;
 fn main() {
     do_things_with_zst::<A>(A);
     do_things_with_zst::<()>(());
-    do_things_with_zst::<[u128; 0]>([]);
+    do_things_with_zst::<[u64; 0]>([]);
     do_things_with_zst::<(A, (A,()))>((A, (A,())));
 }
